puts mina related method in separate class
